bug 1462749: move crashmenow view to django admin
This moves the crashmenow view to the Django admin.

In order to have views that are not Django-model backed but are included
in the admin index page, we override the Django admin index template and
add it manually.
