+++ webapp-django/crashstats/manage/admin.py
@@ -1,0 +1,8 @@
+from crashstats.manage.decorators import superuser_required
+
+
+@superuser_required
+def crash_me_now(request):
+    # NOTE(willkg): This intentionally throws an error so that we can test
+    # unhandled error handling.
+    1 / 0  # noqa
