--- webapp-django/crashstats/manage/jinja2/manage/crash_me_now.html
@@ -1,31 +1,0 @@
-{% extends "manage/base.html" %}
-
-{% block page_title %}{{ super() }} - Crash Me Now{% endblock %}
-
-{% block admin_title %}{{ super() }} - Crash Me Now{% endblock %}
-
-{% block mainbody %}
-<div class="panel">
-    <div class="body notitle">
-
-        <h3>
-            Cause a Webapp Exception
-        </h3>
-
-        <form action="" method="post">
-            {% csrf_token %}
-            <table class="data-table">
-                {{ form }}
-                <tr>
-                    <th>&nbsp;</th>
-                    <td>
-                        <input type="submit" value="Thanks Obama!">
-                    </td>
-                </tr>
-            </table>
-        </form>
-
-    </div>
-</div>
-
-{% endblock %}
