+++ webapp-django/crashstats/manage/jinja2/manage/home.html
@@ -31,10 +31,6 @@
                 <span class="fa fa-check-circle fa-5x"></span>
                 Site Status
             </a></li>
-            <li><a href="{{ url('manage:crash_me_now') }}">
-                <span class="fa fa-bomb fa-5x"></span>
-                Crash Me Now
-            </a></li>
             <li><a href="{{ url('manage:analyze_model_fetches') }}">
                 <span class="fa fa-dashboard fa-5x"></span>
                 Analyze Model Fetches
