+++ webapp-django/crashstats/manage/urls.py
@@ -20,9 +20,6 @@ urlpatterns = [
     url('^supersearch-fields/missing/$',
         views.supersearch_fields_missing,
         name='supersearch_fields_missing'),
-    url('^crash-me-now/$',
-        views.crash_me_now,
-        name='crash_me_now'),
     url('^sitestatus/$',
         views.site_status,
         name='site_status'),
